FitnessAgent/requirements.txt
openai==1.30.0
python-dotenv==1.0.0
requests==2.31.0
duckduckgo-search==5.0.0
dashscope==1.20.0